+++ changelog.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## [unreleased]
 
+- Add css style query capabilities using func Query and ParseExpr
 - files.LoadFunc parses comments aswell
 - Remove cmd/*
 - Remove gregoryv/workdir dependency